+++ src/parsers/graph_parser.cpp
@@ -1,12 +1,12 @@
 #include "graph_parser.h"
 #include <fstream>
 
-std::vector<std::unordered_map<size_t, size_t>> parse_graph(const std::string &graph_path, size_t n) {
-    std::vector<std::unordered_map<size_t, size_t>> edges(n);
+std::vector<std::unordered_map<size_t, double>> parse_graph(const std::string &graph_path, size_t n) {
+    std::vector<std::unordered_map<size_t, double>> edges(n);
     std::ifstream filestream(graph_path);
     size_t start;
     size_t end;
-    size_t weight;
+    double weight;
     while (filestream >> start) {
         filestream >> end;
         filestream >> weight;
