+++ src/parsers/graph_parser.h
@@ -5,6 +5,6 @@
 #include <string>
 #include <unordered_map>
 
-std::vector<std::unordered_map<size_t, size_t>> parse_graph(const std::string &graph_path, size_t n);
+std::vector<std::unordered_map<size_t, double>> parse_graph(const std::string &graph_path, size_t n);
 
 #endif //UNTITLED6_GRAPH_PARSER_H
